test framework: split the P2P transport out of mininode
mininode.py mixed two unrelated concerns: the Zcash consensus data structures
(CTransaction, CBlock, the Sapling and Orchard bundles, and the ser_*/deser_*
primitives) and the Bitcoin P2P half-a-node built on top of them (NodeConn,
NodeConnCB, NetworkThread, and the msg_* wire messages).

Only the P2P half needs asyncore, which PEP 594 removed from the standard
library in Python 3.12, yet its module-level import pinned the entire
repository to Python 3.11: forty test modules import mininode purely for
CTransaction and friends and pay for a dependency they never use.

Move the P2P transport into test_framework/p2p.py and leave mininode.py as the
data structures. p2p.py guards the asyncore import and raises an actionable
error when it is missing. Every test that drives it is already listed in
DISABLED_SCRIPTS, because they exercise zcashd P2P behaviour that zebrad does
not implement, so nothing in the running suite imports it; porting NodeConn and
NetworkThread to asyncio is a prerequisite for re-enabling any of them.

Verified with pyflakes over qa (clean), and by importing every effectively
enabled test script under both Python 3.11 and Python 3.12.
