+++ qa/rpc-tests/invalidtxrequest.py
@@ -7,7 +7,12 @@
 
 from test_framework.test_framework import ComparisonTestFramework
 from test_framework.comptool import TestManager, TestInstance, RejectResult
-from test_framework.mininode import NetworkThread, COIN
+from test_framework.mininode import (
+    COIN,
+)
+from test_framework.p2p import (
+    NetworkThread,
+)
 from test_framework.blocktools import create_block, create_coinbase, create_transaction
 from test_framework.zip317 import MINIMUM_FEE
